Add styling to the add car container
