+++ src/styles/AddCar.css
@@ -1,0 +1,53 @@
+.car-container {
+    display: flex;
+    flex-direction: column;
+    height: 100vh;
+    align-items: center;
+    color: white;
+    justify-content: center;
+    background-image: url('https://source.unsplash.com/random/?lightgreen,abstract');
+    background-size: cover;
+    background-position: center;
+    position: relative;
+    isolation: isolate;
+  }
+  
+  .car-container h1 {
+    padding-bottom: 1rem;
+    border-bottom: 1px solid black;
+  }
+  
+  .car-container::after {
+    content: '';
+    background-color: rgb(120 154 2 / 88%);
+    position: absolute;
+    z-index: -1;
+    inset: 0;
+  }
+  
+  .car-container input,
+  .select-wrapper {
+    background-color: rgba(0, 0, 0, 0.175);
+    border: 1px solid white;
+    color: white;
+    margin-right: 4px;
+  }
+  
+  .car-container input:disabled {
+    background-color: rgba(255, 253, 253, 0.184) !important;
+    border: 2px solid rgba(128, 128, 128, 0.556) !important;
+  }
+  
+  .car-container select {
+    -webkit-appearance: none;
+    appearance: none;
+    background: transparent;
+    border: none;
+    outline: none;
+    color: white;
+  }
+  
+  .car-container select option {
+    color: black;
+  }
+
